+++ database/migrations/2020_10_24_172049_create_dsdcbankdeposits_table.php
@@ -12,8 +12,8 @@ class CreateDsdcbankdepositsTable extends Migration
         Schema::create('dsdcbankdeposits', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('bankaccount_id');
-            $table->float('amount');
-            $table->float('amount_2')->nullable();
+            $table->double('amount');
+            $table->double('amount_2')->nullable();
             $table->string('payment_type');
             $table->string('info');
             $table->date('date');
